refactor: DOM 조작 content_scripts에서 하도록 변경
